requirements.txt
asgiref==3.8.1
Django==5.2.3
django-unfold==0.59.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
